Storing past guesses in a CSV file
I've added functionality to store past guesses in a CSV file. This cuts down the required user inputs down to only 2 inputs (Sequence and actual Guess of current guess) as everything else can be derived from past guesses. Still plenty of improvements to make to this, but it's probably going to be closed out for now as it can consistently get the answer in <= 4 guesses